+++ modelserver/api/__init__.py
@@ -1,0 +1,3 @@
+from .routes import router
+
+__all__ = ["router"]
