+++ modelserver/loaders/core.py
@@ -1,18 +1,13 @@
 from abc import ABC, abstractmethod
-from typing import TypedDict
 
 from modelserver.loaders.locators import Locator
 
-
-class Provenance(TypedDict):
-    """
-    A model can be
-    """
+# Return a new fresh type that can read shit here.
 
 
 class Importer(ABC):
     @abstractmethod
-    def import_from(self, locator: Locator) -> Provenance:
+    def import_from(self, locator: Locator) -> None:
         """
         Accepts a locator that directs the importer to pull the model from the specified location.
 
@@ -24,6 +19,9 @@ class DiskImporter(Importer):
     """
     Imports from local filesystem.
     """
+
+    def import_from(self, locator: Locator) -> None:
+        raise NotImplementedError("unimplemented")
 
 
 # How to decode from JSON directly to the correct locator type.
