+++ modelserver/middleware/__init__.py
@@ -1,0 +1,3 @@
+from .staticfiles import StaticReactRouterFiles
+
+__all__ = ["StaticReactRouterFiles"]
